fix(toolchain): remove known dependency advisories
