+++ apps/web/package.json
@@ -7,7 +7,8 @@
     "dev": "vite",
     "build": "tsc -b && vite build",
     "preview": "vite preview",
-    "lint": "echo \"TODO: add eslint\""
+    "lint": "eslint . --max-warnings=0",
+    "typecheck": "tsc -b --pretty false"
   },
   "dependencies": {
     "@tanstack/react-query": "^5.90.5",
@@ -16,14 +17,20 @@
     "react-router-dom": "^6.30.1"
   },
   "devDependencies": {
+    "@eslint/js": "^10.0.1",
+    "@types/node": "^20.17.32",
     "@types/react": "^18.3.11",
     "@types/react-dom": "^18.3.1",
-    "@types/node": "^20.17.32",
-    "@vitejs/plugin-react": "^4.4.1",
+    "@vitejs/plugin-react": "^6.0.2",
     "autoprefixer": "^10.4.21",
-    "postcss": "^8.5.6",
+    "eslint": "^10.4.0",
+    "eslint-plugin-react-hooks": "^7.1.1",
+    "eslint-plugin-react-refresh": "^0.5.2",
+    "globals": "^17.6.0",
+    "postcss": "^8.5.15",
     "tailwindcss": "^3.4.17",
     "typescript": "^5.9.3",
-    "vite": "^5.4.19"
+    "typescript-eslint": "^8.59.4",
+    "vite": "^8.0.14"
   }
 }
