+++ apps/web/src/components/Badge.tsx
@@ -1,4 +1,4 @@
-type BadgeTone = "neutral" | "good" | "warn" | "bad";
+import type { BadgeTone } from "../lib/badgeTones";
 
 const toneClass: Record<BadgeTone, string> = {
   neutral: "bg-[#1b2838] text-ops-muted",
@@ -9,30 +9,4 @@ const toneClass: Record<BadgeTone, string> = {
 
 export function Badge({ label, tone = "neutral" }: { label: string; tone?: BadgeTone }) {
   return <span className={`inline-flex rounded-full px-2.5 py-1 text-xs font-medium ${toneClass[tone]}`}>{label}</span>;
-}
-
-export function severityTone(severity: string): BadgeTone {
-  switch (severity.toLowerCase()) {
-    case "high":
-      return "bad";
-    case "medium":
-      return "warn";
-    case "low":
-      return "good";
-    default:
-      return "neutral";
-  }
-}
-
-export function recommendationTone(action: string): BadgeTone {
-  switch (action.toLowerCase()) {
-    case "reconcile":
-      return "bad";
-    case "investigate":
-      return "warn";
-    case "monitor":
-      return "good";
-    default:
-      return "neutral";
-  }
 }
