+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/AnouarMohamed/StateSight
 
-go 1.23.0
+go 1.25.10
 
 require (
 	github.com/go-chi/chi/v5 v5.2.3
